move code to period_self.cpp
