+++ toolchain/check/BUILD
@@ -66,6 +66,7 @@ cc_library(
         "operator.cpp",
         "pattern.cpp",
         "pattern_match.cpp",
+        "period_self.cpp",
         "pointer_dereference.cpp",
         "require_impls_stack.cpp",
         "return.cpp",
@@ -135,6 +136,7 @@ cc_library(
         "pattern.h",
         "pattern_match.h",
         "pending_block.h",
+        "period_self.h",
         "pointer_dereference.h",
         "region_stack.h",
         "require_impls_stack.h",
